factor out tweet parsing logic
